refactor(android:layer): use abstract class for XLayer and reduce boilerplate
