+++ android/src/main/java/com/rnmapboxtoolkit/fabric/AbstractMapFeature.kt
@@ -2,8 +2,6 @@ package com.rnmapboxtoolkit.fabric
 
 import android.content.Context
 import com.facebook.react.views.view.ReactViewGroup
-import com.rnmapboxtoolkit.fabric.RnMapboxToolkitView
-
 
 // Source of this file : https://github.com/rnmapbox/maps/blob/b21565f56c858a91ed695650b1ac4a07fcb23098/android/src/main/java/com/rnmapbox/rnmbx/components/AbstractMapFeature.kt
 enum class RemovalReason {
