+++ android/src/main/java/com/rnmapboxtoolkit/fabric/RnMapboxToolkitCircleLayerManager.kt
@@ -1,5 +1,6 @@
 package com.rnmapboxtoolkit.fabric
 
+import com.facebook.react.bridge.Dynamic
 import com.facebook.react.module.annotations.ReactModule
 import com.facebook.react.uimanager.SimpleViewManager
 import com.facebook.react.uimanager.ThemedReactContext
@@ -67,6 +68,13 @@ class RnMapboxToolkitCircleLayerManager :
         value: String?
     ) {
         view?.setLayerStyle(value)
+    }
+
+    override fun setFilter(
+        view: RnMapboxToolkitCircleLayer?,
+        value: Dynamic?
+    ) {
+        view?.setFilter(value)
     }
 
     companion object {
