+++ src/components/lineLayer/LineLayer.type.ts
@@ -1,15 +1,27 @@
+import type { Expression, OnLayerStyleError } from '../../common/shared.types';
+
 export type LineLayerProps = {
   /**
-   * the ID of the layer
+   * The unique identifier for the layer.
    */
   layerID: string;
   /**
-   * the ID of the source
+   * Data source identifier for the layer.
    */
   sourceID: string;
 
-  minzoom?: number;
-  maxzoom?: number;
+  /**
+   * Minimum zoom level at which the layer is visible.
+   * @default undefined (no minimum)
+   */
+  minZoom?: number;
+
+  /**
+   * Maximum zoom level at which the layer is visible.
+   * @default undefined (no maximum)
+   */
+  maxZoom?: number;
+
   layerStyle?: {
     /**
      * The display of line endings.
@@ -185,6 +197,26 @@ export type LineLayerProps = {
      */
     'line-emissive-strength'?: number;
   };
+
+  /**
+   * An expression specifying conditions on source features.
+   * Only features that match the filter are displayed.
+   * Zoom expressions in filters are only evaluated at integer zoom levels
+   */
+  filter?: Expression;
+
+  /**
+     * Event triggered when adding or updating a layer style *properties* fails.
+     *
+     * Note:
+     * - If a style properties is add and fail but not typed in definitions files, this event will be trigger
+     * - Currently, detailed error information may be limited.
+     * @example 
+     * 'Cannot set layer property: circle-radiuss error: circle layer doesn\'t support this property: circle-radiuss' => 'circle-radiuss': 5,
+     * 'Cannot set layer property: circle-color error: value must be a valid color' => 'circle-color': '#fefe',
+     
+     */
+  onLayerStyleError?: (e: OnLayerStyleError) => void;
 };
 
 /**
