+++ src/specs/NativeComponentCircleLayer.ts
@@ -7,6 +7,8 @@ import type {
   Double,
   DirectEventHandler,
 } from 'react-native/Libraries/Types/CodegenTypesNamespace';
+import type { Expression } from '../common/shared.types';
+import type { UnsafeMixed } from './codegenUtils';
 
 type OnLayerStyleErrorEventHandler = DirectEventHandler<
   Readonly<{ properties: { message: string } }>
@@ -22,6 +24,8 @@ interface NativeComponentsCircleLayerProps extends ViewProps {
    * Omit type is refused by codegen so use other name only from ts -> kt
    */
   layerStyle?: string;
+
+  filter?: UnsafeMixed<Expression>;
   onLayerStyleError?: OnLayerStyleErrorEventHandler;
 }
