+++ src/specs/NativeComponentLineLayer.ts
@@ -3,11 +3,12 @@ import {
   type HostComponent,
   type ViewProps,
 } from 'react-native';
-
 import type {
-  Double,
   DirectEventHandler,
+  Double,
 } from 'react-native/Libraries/Types/CodegenTypesNamespace';
+import type { UnsafeMixed } from './codegenUtils';
+import type { Expression } from '../common/shared.types';
 
 type OnLayerStyleErrorEventHandler = DirectEventHandler<
   Readonly<{ properties: { message: string } }>
@@ -22,6 +23,8 @@ interface NativeComponentsLineLayerProps extends ViewProps {
    * Omit type is refused by codegen so use other name only from ts -> kt
    */
   layerStyle?: string;
+
+  filter?: UnsafeMixed<Expression>;
 
   onLayerStyleError?: OnLayerStyleErrorEventHandler;
 }
